+++ src/ChessDotCpp/options.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "types.h"
+
+class Options
+{
+public:
+
+	static inline ThreadId Threads = 1;
+	
+	Options() = delete;
+};