+++ src/ChessDotCpp/types.h
@@ -26,5 +26,6 @@ using Magic = uint64_t;
 using Time = int32_t;
 using Stat = uint64_t;
 using TtFlag = uint8_t;
+using ThreadId = int8_t;
 
 using Fen = std::string;